se actualiza el codigo como un putas fin

+++ src/main/java/com/cloud/parcial/repositories/ProjectRepository.java
@@ -1,8 +1,7 @@
 package com.cloud.parcial.repositories;
 
-import com.cloud.parcial.intities.BacklogEntity;
 import com.cloud.parcial.intities.ProjectEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProjectRepository extends CrudRepository<ProjectEntity, Long> {
+public interface ProjectRepository extends CrudRepository<ProjectEntity, Integer> {
 }
