+++ src/main/java/com/cloud/parcial/repositories/ProjectTaskRepository.java
@@ -3,6 +3,6 @@ package com.cloud.parcial.repositories;
 import com.cloud.parcial.intities.ProjectTaskEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProjectTaskRepository extends CrudRepository<ProjectTaskEntity, Long> {
+public interface ProjectTaskRepository extends CrudRepository<ProjectTaskEntity, Integer> {
 
 }
